RAG-LLM-for-Ports-backend/requirements.txt
# AI Port Decision-Support System - Requirements
# Complete dependency list for production deployment

# Core LangChain and AI
langchain>=0.1.0,<1.0.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
langgraph>=0.2.0

# Vector Database and Embeddings
chromadb>=0.4.18
sentence-transformers>=2.2.2
openai>=1.3.7

# Document Processing
PyPDF2>=3.0.1
pypdf>=3.17.4
unstructured>=0.11.6
python-docx>=1.1.0

# Multilingual Support
transformers>=4.36.2
torch>=2.1.1,<2.6.0
langdetect>=1.0.9
tiktoken>=0.5.2
accelerate>=0.24.0
safetensors>=0.4.0
huggingface-hub>=0.24.0

# OCR Support for Scanned Documents
pytesseract>=0.3.10
easyocr>=1.7.0
opencv-python>=4.8.1.78
Pillow>=10.1.0
PyMuPDF>=1.23.0

# Web Framework and API
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Data Processing
pandas>=2.1.4
numpy>=1.24.3,<2.0.0
python-dateutil>=2.8.2

# Database and Storage
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9

# Security and Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.2

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiofiles>=23.2.1
httpx>=0.25.2

# Development and Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
flake8>=6.1.0

# Monitoring and Logging
structlog>=23.2.0
loguru>=0.7.2